+++ src/main/java/fll/xml/ui/NonNumericCategoryEditor.java
@@ -31,15 +31,13 @@ import fll.xml.NonNumericCategory;
 
   private static final org.apache.logging.log4j.Logger LOGGER = org.apache.logging.log4j.LogManager.getLogger();
 
-  private final JFormattedTextField nameEditor;
-
   private final JFormattedTextField titleEditor;
 
   private final JCheckBox perAwardGroupEditor;
 
   private final TextAreaEditor description;
 
-  private final ValidityPanel validPanel;
+  private final ValidityPanel validPanel;  
 
   /* package */ NonNumericCategoryEditor(final NonNumericCategory category) {
     super(new GridBagLayout());
@@ -59,24 +57,6 @@ import fll.xml.NonNumericCategory;
     gbc.weightx = 0;
     gbc.anchor = GridBagConstraints.FIRST_LINE_END;
     add(new JLabel("Name: "), gbc);
-
-    nameEditor = FormatterUtils.createDatabaseNameField();
-    gbc = new GridBagConstraints();
-    gbc.weightx = 1;
-    gbc.anchor = GridBagConstraints.FIRST_LINE_START;
-    gbc.gridwidth = GridBagConstraints.REMAINDER;
-    gbc.fill = GridBagConstraints.HORIZONTAL;
-    add(nameEditor, gbc);
-
-    nameEditor.setColumns(40);
-    nameEditor.setMaximumSize(nameEditor.getPreferredSize());
-
-    nameEditor.addPropertyChangeListener("value", e -> {
-      if (null != this.category) {
-        final String newName = nameEditor.getText();
-        this.category.setName(newName);
-      }
-    });
 
     gbc = new GridBagConstraints();
     gbc.weightx = 0;
@@ -155,12 +135,6 @@ import fll.xml.NonNumericCategory;
    */
   public void commitChanges() {
     try {
-      nameEditor.commitEdit();
-    } catch (final ParseException e) {
-      LOGGER.debug("Got parse exception committing changes to name, assuming bad value and ignoring", e);
-    }
-
-    try {
       titleEditor.commitEdit();
     } catch (final ParseException e) {
       LOGGER.debug("Got parse exception committing changes to title, assuming bad value and ignoring", e);
@@ -173,10 +147,6 @@ import fll.xml.NonNumericCategory;
   @Override
   public boolean checkValidity(final Collection<String> messagesToDisplay) {
     final List<String> messages = new LinkedList<>();
-
-    if (StringUtils.isBlank(nameEditor.getText())) {
-      messages.add("The category must have a name");
-    }
 
     if (StringUtils.isBlank(titleEditor.getText())) {
       messages.add("The non-numeric category must have a title");